chore: Beautify report layout
